requirements.txt
yfinance
kafka-python
cassandra-driver